+++ src/lib.rs
@@ -40,7 +40,6 @@ use miniquad::*;
 
 use std::collections::{HashMap, HashSet};
 use std::future::Future;
-use std::mem::ManuallyDrop;
 use std::pin::Pin;
 
 mod exec;
@@ -453,7 +452,7 @@ impl Context {
 }
 
 #[no_mangle]
-static mut CONTEXT: ManuallyDrop<Option<Context>> = ManuallyDrop::new(None);
+static mut CONTEXT: Option<Context> = None;
 
 // This is required for #[macroquad::test]
 //
@@ -844,7 +843,7 @@ impl Window {
             }
             let mut context = Context::new();
             context.update_on = update_on.unwrap_or_default();
-            unsafe { CONTEXT = ManuallyDrop::new(Some(context)) };
+            unsafe { CONTEXT = Some(context) };
 
             Box::new(Stage {})
         });
