+++ tests/coroutine_values.rs
@@ -1,7 +1,7 @@
-use macroquad::{experimental::coroutines::start_coroutine, telemetry, window::next_frame};
+use macroquad::{experimental::coroutines::start_coroutine, window::next_frame};
 
 #[macroquad::test]
-async fn coroutine_value() {
+async fn coroutine_values() {
     let mut coroutine = start_coroutine(async move {
         next_frame().await;
         1
@@ -15,22 +15,4 @@ async fn coroutine_value() {
     coroutine.poll(0.0);
 
     assert_eq!(coroutine.retrieve(), Some(1));
-}
-
-#[macroquad::test]
-async fn coroutine_memory() {
-    use macroquad::prelude::*;
-
-    for _ in 0..20 {
-        start_coroutine(async move {
-            next_frame().await;
-        });
-
-        next_frame().await;
-    }
-
-    // wait for the last one to finish
-    next_frame().await;
-
-    assert_eq!(telemetry::active_coroutines_count(), 0);
 }
